refactor block list to create better separation of concerns, break up excessively large single component
